refactor: move favicon.go to v1/favicon
