+++ internal/api/v1/favicon/favicon.go
@@ -1,4 +1,4 @@
-package v1
+package favicon
 
 import (
 	"errors"
